# frontend/public/api/openapi.yaml
openapi: 3.0.3
info:
  title: Nebula DW API
  version: 1.0.0
  description: API for managing nebula data warehouse operations, including bikes, customers, projects, and role management.
servers:
  - url: https://api.example.com
paths:
...
  /customers:
    get:
      summary: Retrieve a list of customers
      operationId: getCustomers
      parameters:
        - $ref: "#/components/parameters/FilterParam"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/OrderBy"
      responses:
        '200':
          description: A list of customers with total count
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCount:
                    type: integer
                    example: 42
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – insufficient permissions for projects
        '500':
          description: Internal server error
      tags:
        - Customers
    delete:
      summary: Delete a customer by ID
      operationId: deleteCustomer
      parameters:
        - $ref: "#/components/parameters/DeleteIdParam"
        - $ref: "#/components/parameters/DeleteCascadeParam"
      responses:
        '204':
          description: Customer deleted successfully
        '409':
          description: Conflict – related data exists; use cascade=true to force delete
        '400':
          description: Bad request – missing or invalid ID
        '401':
          description: Unauthorized – missing or invalid token
        '500':
          description: Internal server error
      tags:
        - Customers
    put:
      summary: Update a customer
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer updated successfully
        '400':
          description: Bad request – missing or invalid fields
        '401':
          description: Unauthorized – missing or invalid token
        '404':
          description: Customer not found
        '500':
          description: Internal server error
      tags:
        - Customers
    post:
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created successfully
        '400':
          description: Bad request – missing or invalid fields
        '401':
          description: Unauthorized – missing or invalid token
        '409':
          description: Conflict – duplicate or invalid data
        '500':
          description: Internal server error
      tags:
        - Customers
...
  /rolemanagements:
    get:
      summary: Retrieve a list of role managements with project names
      operationId: getRoleManagements
      parameters:
        - $ref: "#/components/parameters/FilterParam"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/OrderBy"
      responses:
        '200':
          description: A list of role managements with project names with total count
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCount:
                    type: integer
                    example: 42
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RoleManagementWithName'
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – insufficient permissions for projects
        '500':
          description: Internal server error
      tags:
        - RoleManagements
    delete:
      summary: Delete a role management by email and project ID
      operationId: deleteRoleManagementByEmailAndProjectId
      parameters:
        - $ref: "#/components/parameters/DeleteEmailParam"
        - $ref: "#/components/parameters/DeleteProjectIdParam"
      responses:
        '204':
          description: Role management deleted successfully
        '400':
          description: Bad request – missing or invalid ID
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – with current permissions not allowed to delete
        '500':
          description: Internal server error
      tags:
        - RoleManagements
    put:
      summary: Update a role management
      operationId: updateRoleManagement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleManagement'
      responses:
        '200':
          description: Role management updated successfully
        '400':
          description: Bad request – missing or invalid fields
        '401':
          description: Unauthorized – missing or invalid token
        '404':
          description: Role management not found
        '500':
          description: Internal server error
      tags:
        - RoleManagements
    post:
      summary: Create a new role management
      operationId: createRoleManagement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleManagement'
      responses:
        '201':
          description: Role management created successfully
        '400':
          description: Bad request – missing or invalid fields
        '401':
          description: Unauthorized – missing or invalid token
        '409':
          description: Conflict – duplicate or invalid data
        '500':
          description: Internal server error
      tags:
        - RoleManagements
  /rolemanagements/{id}:
    get:
      summary: Retrieve a list of role managements for a given ID
      operationId: getRoleManagementsById
      parameters:
        - $ref: "#/components/parameters/GetIdParam"
      responses:
        '200':
          description: A list of role management entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleManagement'
        '401':
          description: Unauthorized – missing or invalid token
        '403':
          description: Forbidden – insufficient permissions
        '404':
          description: Not Found – no role managements for the given ID
        '500':
          description: Internal server error
      tags:
        - RoleManagements
...
components:
  parameters:
    PageSize:
      name: pageSize
      in: query
      description: Specifying the size of the result set.
      required: false
      allowEmptyValue: false
      example: 100
      schema:
        type: integer
        format: integer
        minimum: 1
        maximum: 500
    PageParam:
      name: page
      in: query
      description: Specifying the page of the result set.
      required: false
      allowEmptyValue: false
      example: 0 or 1
      schema:
        type: integer
        format: integer
        minimum: 0
    FilterParam:
      name: filter
      in: query
      description: Query filter string, e.g. project_id:$eq.1|2|3
      required: false
      schema:
        type: string
    OrderBy:
      name: orderBy
      in: query
      description: Specifying the sort order. The sort definition can use the keywords 'asc' for ascending and 'desc' for descending sort order.
      required: false
      allowEmptyValue: false
      example: id desc
      schema:
        type: string
        description: Comma seperated list of sort definitions
    GetIdParam:
      name: id
      in: path
      required: true
      description: ID of the primary key to retrieve
      schema:
        type: integer
        example: 123
    DeleteIdParam:
      name: id
      in: query
      description: ID of the primary key to delete
      required: true
      schema:
        type: integer
        example: 123
    DeleteProjectIdParam:
      name: project_id
      in: query
      description: Project ID of the primary key to delete
      required: true
      schema:
        type: integer
        example: 123
    DeleteEmailParam:
      name: email
      in: query
      description: Email address of the user to retrieve or modify
      required: true
      schema:
        type: string
        format: email
        example: user@example.com
    DeleteCascadeParam:
      name: cascade
      in: query
      description: Whether to delete related data as well
      required: false
      schema:
        type: boolean
        example: true
  schemas:
...
    RoleManagement:
      required: [ user_email, project_id, role ]
      type: object
      properties:
        user_email:
          type: string
          description: Email of the user
        project_id:
          type: integer
          description: Associated project ID
        role:
          type: string
          description: Role assigned to the user in the project
    Customer:
      type: object
      required: [ email, password, first_name, name, dob, city, project_id ]
      properties:
        id:
          type: integer
          description: Unique identifier for the customer
        email:
          type: string
          description: Customer's email address
        password:
          type: string
          description: Hashed password for authentication
        first_name:
          type: string
          description: Customer's given name
        name:
          type: string
          description: Customer's family name
        dob:
          type: string
          format: date
          description: Date of birth (YYYY-MM-DD)
        city:
          type: string
          description: City of residence
        project_id:
          type: integer
          description: Project ID associated with the customer
...
    RoleManagementWithName:
      allOf:
        - $ref: '#/components/schemas/RoleManagement'
        - type: object
          required: [ project_name ]
          properties:
            project_name:
              type: string
              description: Name of the associated project
...